some test refactoring/DRY work
though I often feel DRY work is premature and not helpful, I was
absolutely forgetting to test that all my auth'ed routes had the auth
decorator applied. I wanted a general way to test all of that and added
some more meat to the test parent class to do that.

I'm not thrilled with what I did (especially since pytest picks up the
generalized tests in the parent test class) but it works.